requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[argon2]
python-jose
pydantic
python-multipart
aiofiles